Move mainwindow code to gui module so that other modules can have access to it.  Add screen changing behavior.
The mainWindow object and the screens array 
really needed to be accessible to other parts
of the program.  All those signals that were in
the hooks.py file that seemed so clumsy were
really there to get around the fact that the main
window was in the __main__ module.  Now other modules
can connect to the signals in mainWindow directly
as well as call functions.