Fix a bug that happened when the same screen was selected twice.
Because the hide() call is after the show() call if the same
screen is selected to be shown that is already being shown
then no screens were shown.  Don’t know if the hide call 
needs to be second but it seems like it’d be a cleaner update
that way.