Specify config path relative to main.py
As of now, the path to the config directory is specified in an absolute
manner that works when you run pyEfis from the project root. For the
distributed version, however, the path is different. Therefore, main.py
will now assemble the path based on the path of the main.py file.